agent-system-b/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
pydantic>=2.9,<3.0
pydantic-settings>=2.5,<3.0
orjson>=3.10,<4.0
google-adk>=1.0,<2.0
